rename xml version
